add exceptions

+++ src/main/java/com/employee/controller/AddressController.java
@@ -1,6 +1,7 @@
 package com.employee.controller;
 
 import com.employee.dto.AddressDTO;
+import com.employee.exception.AddressException;
 import com.employee.model.Address;
 import com.employee.objectsMappers.Mapper;
 import com.employee.service.GenericService;
@@ -8,6 +9,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/addresses")
@@ -19,34 +22,52 @@ public class AddressController {
 
     @GetMapping("/{id}")
     public ResponseEntity<AddressDTO> getAddressById(@PathVariable Long id) {
-        Address address = addressService.getById(id);
-        AddressDTO addressDTO = addressMapper.mapToDTO(address);
-        return ResponseEntity.ok(addressDTO);
+        Optional<Address> addressOptional = addressService.getById(id);
+        if (addressOptional.isPresent()) {
+            Address address = addressOptional.get();
+            AddressDTO addressDTO = addressMapper.mapToDTO(address);
+            return ResponseEntity.ok(addressDTO);
+        } else {
+            throw new AddressException("Address with ID " + id + " not found.", HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping
     public ResponseEntity<AddressDTO> createAddress(@RequestBody AddressDTO addressDTO) {
         Address address = addressMapper.mapFromDTO(addressDTO);
-        Address savedAddress = addressService.save(address);
-        AddressDTO savedAddressDTO = addressMapper.mapToDTO(savedAddress);
-        return ResponseEntity.status(HttpStatus.CREATED).body(savedAddressDTO);
+        Optional<Address> savedAddressOptional = addressService.save(address);
+        if (savedAddressOptional.isPresent()) {
+            Address savedAddress = savedAddressOptional.get();
+            AddressDTO savedAddressDTO = addressMapper.mapToDTO(savedAddress);
+            return ResponseEntity.status(HttpStatus.CREATED).body(savedAddressDTO);
+        } else {
+            throw new AddressException("Failed to create address.", HttpStatus.BAD_REQUEST);
+        }
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<AddressDTO> updateAddress(@PathVariable Long id, @RequestBody AddressDTO addressDTO) {
-        Address existingAddress = addressService.getById(id);
+        Address existingAddress = addressService.getById(id)
+                .orElseThrow(() -> new AddressException("Address with ID " + id + " not found.", HttpStatus.NOT_FOUND));
+
         Address updatedAddress = addressMapper.mapFromDTO(addressDTO);
         updatedAddress.setId(existingAddress.getId());
-        Address savedAddress = addressService.update(updatedAddress);
+
+        Optional<Address> savedAddressOptional = addressService.update(updatedAddress);
+        Address savedAddress = savedAddressOptional.get();
         AddressDTO savedAddressDTO = addressMapper.mapToDTO(savedAddress);
         return ResponseEntity.ok(savedAddressDTO);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteAddress(@PathVariable Long id) {
-        Address address = addressService.getById(id);
-        addressService.delete(address);
-        return ResponseEntity.noContent().build();
+        Optional<Address> addressOptional = addressService.getById(id);
+        if (addressOptional.isPresent()) {
+            Address address = addressOptional.get();
+            addressService.delete(address);
+            return ResponseEntity.noContent().build();
+        } else {
+            throw new AddressException("Address with ID " + id + " not found.", HttpStatus.NOT_FOUND);
+        }
     }
 }
-
